+++ src/js/components/skin-badge.js
@@ -9,6 +9,7 @@
  *
  * Optional attributes:
  *   - data-icon (Any valid skin-icon name)
+ *   - data-style (Alternate style: pill)
  */
 export class SkinBadge extends HTMLElement {
     constructor() {
@@ -17,19 +18,17 @@ export class SkinBadge extends HTMLElement {
 
     connectedCallback() {
 
-        const badgeContents = this.innerHTML;
+        const contents = this.innerHTML;
 
-        this.classList.add("tc-badge");
-
-        let badgeHtml = '';
+        let innerHtml = '';
 
         if (typeof this.dataset.icon != "undefined") {
-            badgeHtml += '<skin-icon name="' + this.dataset.icon + '" class="flex-none pt-0.5 pr-1" data-class="w-4 h-4"></skin-icon>';
+            innerHtml += '<skin-icon name="' + this.dataset.icon + '" class="flex-none pt-0.5 pr-1" data-class="w-4 h-4"></skin-icon>';
         }
 
-        badgeHtml += badgeContents;
+        innerHtml += contents;
 
-        this.innerHTML = badgeHtml;
+        this.innerHTML = innerHtml;
 
     }
 }